@@ -13,11 +13,11 @@
     "check": "tsc --noEmit"
   },
   "dependencies": {
-    "@wundergraph/nextjs": "^0.7.0",
-    "@wundergraph/sdk": "^0.125.0",
+    "@wundergraph/nextjs": "^0.7.7",
+    "@wundergraph/sdk": "^0.130.2",
     "@wundergraph/swr": "^0.2.2",
     "graphql": "^16.3.0",
-    "next": "^12.1.6",
+    "next": "^13.1.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "swr": "^2.0.0"
