@@ -15,10 +15,10 @@
   "license": "ISC",
   "dependencies": {
     "@apollo/client": "^3.6.9",
-    "@wundergraph/nextjs": "^0.7.0",
-    "@wundergraph/sdk": "^0.125.0",
+    "@wundergraph/nextjs": "^0.7.7",
+    "@wundergraph/sdk": "^0.130.2",
     "graphql": "^16.5.0",
-    "next": "^12.1.6",
+    "next": "^13.1.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "swr": "^2.0.0"
