@@ -20,27 +20,27 @@
     "test": "jest"
   },
   "dependencies": {
-    "@wundergraph/nextjs": "^0.7.0",
-    "@wundergraph/sdk": "^0.125.0",
+    "@wundergraph/nextjs": "^0.7.7",
+    "@wundergraph/sdk": "^0.130.2",
     "graphql": "^16.3.0",
-    "next": "^12.1.6",
+    "next": "^13.1.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "swr": "^2.0.0"
   },
   "devDependencies": {
+    "@jest/globals": "^29.3.1",
     "@types/node": "^17.0.15",
-    "@types/react": "^18.0.6",
     "@types/node-fetch": "^2.6.2",
+    "@types/react": "^18.0.6",
+    "jest": "^29.3.1",
     "node-fetch": "^2.6.7",
     "npm-run-all": "^4.1.5",
     "open-cli": "^7.0.1",
     "prisma": "^4.0.0",
+    "ts-jest": "^29.0.3",
     "ts-node": "^10.8.0",
     "typescript": "^4.8.2",
-    "wait-on": "^6.0.1",
-    "@jest/globals": "^29.3.1",
-    "jest": "^29.3.1",
-    "ts-jest": "^29.0.3"
+    "wait-on": "^6.0.1"
   }
 }
