@@ -1,4 +1,4 @@
-import { configureWunderGraphServer } from '@wundergraph/sdk';
+import { configureWunderGraphServer } from '@wundergraph/sdk/server';
 import type { HooksConfig } from './generated/wundergraph.hooks';
 import type { InternalClient } from './generated/wundergraph.internal.client';
 
