@@ -1,4 +1,4 @@
-import type { Webhook, WebhookHttpEvent, WebhookHttpResponse } from '@wundergraph/sdk';
+import type { Webhook, WebhookHttpEvent, WebhookHttpResponse } from '@wundergraph/sdk/server';
 import type { InternalClient } from '../generated/wundergraph.internal.client';
 
 const webhook: Webhook<
