@@ -1,6 +1,6 @@
 //language=handlebars
 export const template = `
-import { WebhookConfiguration } from '@wundergraph/sdk';
+import type { WebhookConfiguration } from '@wundergraph/sdk/server';
 
 export type WebhooksConfig = {
 {{#each webhooks}}
