@@ -1,4 +1,4 @@
-export type { ResolvedWunderGraphConfig, HooksConfiguration, OperationHooksConfiguration } from './configure';
+export type { ResolvedWunderGraphConfig } from './configure';
 export type {
 	BaseOperationConfiguration,
 	QueryConfiguration,
@@ -13,67 +13,15 @@ export type {
 	CustomizeSubscription,
 } from './configure/operations';
 export type { GraphQLOperation } from './graphql/operations';
-export type {
-	WunderGraphRequest,
-	WunderGraphResponse,
-	WsTransportOnConnectionInitResponse,
-	WunderGraphUser,
-	BaseRequestContext,
-	AuthenticationRequestContext,
-	AuthenticationResponse,
-	ClientRequest,
-	WunderGraphHooksAndServerConfig,
-	WunderGraphServerConfig,
-	ClientRequestHeaders,
-	AuthenticationHookRequest,
-} from './middleware/types';
-export type {
-	SubscriptionRequestOptions,
-	Response,
-	ResponseOK,
-	RequestOptions,
-	MutateRequestOptions,
-	CachedResponse,
-	UploadResponse,
-	ClientConfig,
-	RequiresAuthentication,
-	Error,
-	Headers,
-	Lazy,
-	None,
-	Loading,
-	Refetch,
-	UserListener,
-	Aborted,
-	FetchConfig,
-} from './middleware/client';
-export type { OperationArgsWithInput, InternalClient } from './middleware/internal-client';
-export type { GraphQLServerConfig } from './middleware/plugins/graphql';
-export type { FastifyLoggerInstance } from 'fastify';
 export type { Template, TemplateOutputFile } from './codegen/index';
-export type {
-	WebhookHttpEvent,
-	WebhookHttpResponse,
-	Webhook,
-	WebhookRequestContext,
-	WebhookLogger,
-	WebhookHeaders,
-	WebhookQuery,
-	WebhookConfiguration,
-} from './webhooks/types';
-export type { WebhookVerifierConfiguration } from './webhooks/verifiers';
-
 export { default as templates } from './codegen/templates';
 export { introspect, createMockApi, Api } from './definition';
 export { configureWunderGraphApplication } from './configure';
 export { configureWunderGraphOperations, enableAuth, enableCaching, disableAuth } from './configure/operations';
-export { configureWunderGraphServer } from './middleware/server';
-
 export { default as cors } from './cors';
 export { authProviders } from './configure/authentication';
 export type { LoggerLevel } from './configure/options';
 export { WgEnv } from './configure/options';
-
 export {
 	BaseTypeScriptDataModel,
 	formatTypeScript,
@@ -82,7 +30,6 @@ export {
 	TypeScriptResponseModels,
 } from './codegen/templates/typescript/index';
 export { TypeScriptClient } from './codegen/templates/typescript/client';
-export { GithubWebhookVerifier, CreateWebhookVerifier, WebhookVerifierKind } from './webhooks/verifiers';
 export { PlaceHolder } from './configure/variables';
 export { EnvironmentVariable } from './configure/variables';
 
