@@ -2,9 +2,9 @@ import { FastifyPluginAsync } from 'fastify';
 import path from 'path';
 import { Webhook, WebhookHeaders, WebhookQuery } from '../../webhooks/types';
 import { Headers } from '@web-std/fetch';
-import { WebhookConfiguration } from '@wundergraph/protobuf';
-import { InternalClientFactory } from '../internal-client';
-import { RequestMethod } from '../types';
+import type { RequestMethod } from '../types';
+import type { WebhookConfiguration } from '@wundergraph/protobuf';
+import type { InternalClientFactory } from '../internal-client';
 
 export interface WebHookRouteConfig {
 	kind: 'webhook';
