@@ -1,19 +1,24 @@
-import { WunderGraphConfiguration } from '@wundergraph/protobuf';
 import closeWithGrace from 'close-with-grace';
 import { Headers } from '@web-std/fetch';
 import process from 'node:process';
-import HooksPlugin, { HooksRouteConfig } from './plugins/hooks';
-import FastifyWebhooksPlugin, { WebHookRouteConfig } from './plugins/webhooks';
-import GraphQLServerPlugin from './plugins/graphql';
 import Fastify, { FastifyInstance } from 'fastify';
-import { customGqlServerMountPath, HooksConfiguration } from '../configure';
 import type { InternalClient } from './internal-client';
 import { InternalClientFactory, internalClientFactory } from './internal-client';
 import { pino } from 'pino';
 import path from 'path';
 import fs from 'fs';
-import {
+import { ServerLogger, resolveServerLogLevel } from '../logger';
+import { resolveConfigurationVariable } from '../configure/variables';
+import { onParentProcessExit } from '../utils/process';
+import { customGqlServerMountPath } from './util';
+
+import type { WunderGraphConfiguration } from '@wundergraph/protobuf';
+import type { WebhooksConfig } from '../webhooks/types';
+import type { HooksRouteConfig } from './plugins/hooks';
+import type { WebHookRouteConfig } from './plugins/webhooks';
+import type {
 	FastifyRequestBody,
+	HooksConfiguration,
 	ServerRunOptions,
 	WunderGraphHooksAndServerConfig,
 	WunderGraphServerConfig,
@@ -89,12 +94,13 @@ const _configureWunderGraphServer = <
 	/**
 	 * Configure the custom GraphQL servers
 	 */
-	if (serverConfig.graphqlServers) {
+	if (serverConfig.graphqlServers && serverConfig.graphqlServers.length > 0) {
 		let seenServer: { [key: string]: boolean } = {};
 		serverConfig.graphqlServers.forEach((server) => {
 			if (seenServer[server.serverName]) {
 				throw new Error(`A server with the name '${server.serverName}' has been already registered!`);
 			}
+			server.routeUrl = customGqlServerMountPath(server.serverName);
 			seenServer[server.serverName] = true;
 		});
 	}
@@ -215,23 +221,23 @@ export const createServer = async ({
 			};
 		});
 
-		if (serverConfig?.hooks) {
-			await fastify.register(HooksPlugin, { ...serverConfig.hooks, config });
+		if (serverConfig?.hooks && Object.keys(serverConfig.hooks).length > 0) {
+			await fastify.register(require('./plugins/hooks'), { ...serverConfig.hooks, config });
 			fastify.log.info('Hooks plugin registered');
 		}
 
-		if (serverConfig.graphqlServers) {
+		if (serverConfig.graphqlServers && serverConfig.graphqlServers.length > 0) {
 			for await (const server of serverConfig.graphqlServers) {
 				const routeUrl = customGqlServerMountPath(server.serverName);
-				await fastify.register(GraphQLServerPlugin, { ...server, routeUrl: routeUrl });
+				await fastify.register(require('./plugins/graphql'), { ...server, routeUrl: routeUrl });
 				fastify.log.info('GraphQL plugin registered');
 				fastify.log.info(`Graphql server '${server.serverName}' listening at ${routeUrl}`);
 			}
 		}
 	});
 
-	if (config.api) {
-		await fastify.register(FastifyWebhooksPlugin, {
+	if (config.api?.webhooks && config.api.webhooks.length > 0) {
+		await fastify.register(require('./plugins/webhooks'), {
 			wundergraphDir,
 			webhooks: config.api.webhooks,
 			internalClientFactory: clientFactory,
