@@ -202,10 +202,10 @@ export class WunderGraphTestServer<ClientType extends Client = Client> {
 			manageServer: true,
 			env: {
 				WG_CLOUD: 'true',
-				WG_NODE_URL: `http://localhost:${nodePort}`,
+				WG_NODE_URL: `http://0.0.0.0:${nodePort}`,
 				WG_NODE_HOST: '0.0.0.0',
 				WG_NODE_PORT: nodePort.toString(),
-				WG_SERVER_URL: `http://localhost:${serverPort}`,
+				WG_SERVER_URL: `http://0.0.0.0:${serverPort}`,
 				WG_SERVER_HOST: '0.0.0.0',
 				WG_SERVER_PORT: serverPort.toString(),
 			},
