@@ -5,16 +5,24 @@ import * as fs from 'fs';
 import rimraf from 'rimraf';
 import { logger } from './logger';
 import path from 'path';
+import type debug from 'debug';
 
 export const installer = async (version: string, installDir: string, binaryName: string) => {
 	const log = logger.extend('install');
 	const error = logger.extend('error:install');
 
 	const lockFile = path.join(installDir, 'download-lock');
+	const binaryPath = path.join(installDir, binaryName);
 	const locker = LockFile(version, lockFile);
 
 	if (locker.exists()) {
-		log(`Lock file already exists, skipping the download of the binary ${version}`);
+		log(`Lock file already exists, skipping the download of the binary v${version}`);
+		// That's a convenience, so we have a fixed path to current installed binary
+		if (process.env.WG_COPY_BIN_PATH) {
+			log(`copy v${version} binary to ${process.env.WG_COPY_BIN_PATH}`);
+			copyFileRecursive(log, binaryPath, process.env.WG_COPY_BIN_PATH);
+			log(`binary copied`);
+		}
 		return;
 	}
 
@@ -24,7 +32,6 @@ export const installer = async (version: string, installDir: string, binaryName:
 
 	locker.create();
 
-	const binaryPath = path.join(installDir, binaryName);
 	if (fs.existsSync(binaryPath)) {
 		log(`removing existing binary at ${binaryPath}`);
 		rimraf(binaryPath, (err) => {
@@ -48,6 +55,13 @@ export const installer = async (version: string, installDir: string, binaryName:
 			log(`wunderctl v${version} installed/updated`);
 			log(`make binary executable`);
 			chmodX(binaryPath);
+
+			// That's a convenience, so we have a fixed path to current installed binary
+			if (process.env.WG_COPY_BIN_PATH) {
+				log(`copy v${version} binary to ${process.env.WG_COPY_BIN_PATH}`);
+				copyFileRecursive(log, binaryPath, process.env.WG_COPY_BIN_PATH);
+				log(`binary copied`);
+			}
 		});
 		outStream.addListener('error', (err) => {
 			error('Error installing wunderctl: ' + err.message);
@@ -58,6 +72,11 @@ export const installer = async (version: string, installDir: string, binaryName:
 		error('Error installing wunderctl: ' + e.message);
 	}
 };
+
+function copyFileRecursive(log: debug.Debugger, sourceFilePath: string, targetFilePath: string) {
+	fs.mkdirSync(path.dirname(targetFilePath), { recursive: true });
+	fs.copyFileSync(sourceFilePath, targetFilePath);
+}
 
 function LockFile(version: string, lockFile: string) {
 	let createdLockFile = false;
@@ -79,10 +98,7 @@ function LockFile(version: string, lockFile: string) {
 		exists: () => {
 			if (fs.existsSync(lockFile)) {
 				const data = fs.readFileSync(lockFile, 'utf-8');
-				if (data === version) {
-					return true;
-				}
-				return false;
+				return data === version;
 			}
 		},
 	};
